Refactor visualization functions.
